+++ src/constant/query/GOOGLE_PARAMS.ts
@@ -1,0 +1,11 @@
+const GOOGLE_PARAMS = {
+  SIZE: "isz",
+  COLOR: "ic",
+  SPECIFIC_COLOR: "Cisc",
+  TYPE: "itp",
+  DATE: "qdr",
+  LICENCE: "il",
+  IMAGE_EXTENSION: "ift",
+};
+
+export default GOOGLE_PARAMS;