+++ types/results.d.ts
@@ -1,0 +1,8 @@
+import ImageResultItem from "./imageResultItem";
+
+type Results = {
+  url: string;
+  result: ImageResultItem[];
+};
+
+export default Results;